SmartSDLC/requirements.txt
fastapi
uvicorn
transformers
huggingface_hub
torch
PyMuPDF
streamlit
requests
streamlit-lottie
python-dotenv
